Refactoring for ggez example
