+++ examples/ggez/main.rs
@@ -1,29 +1,32 @@
-use std::{collections::HashMap, path::PathBuf, sync::Arc};
+mod map;
 
-use ggez::{event::{self, MouseButton}, Context, GameResult, graphics::{self, spritebatch::SpriteBatch, DrawParam}, input};
-use tiled::{Map, LayerData, FilesystemResourceCache, TileLayer, Tileset};
+use ggez::{event::{self, MouseButton}, Context, GameResult, graphics::{self, DrawParam}, input};
+use map::MapHandler;
+use tiled::{Map, FilesystemResourceCache};
 
 fn main() -> GameResult {
-    let cb = ggez::ContextBuilder::new("tiled + ggez", "tiled")
+    // init ggez
+    let cb = ggez::ContextBuilder::new("rs-tiled + ggez", "rs-tiled")
         .window_setup(ggez::conf::WindowSetup::default()
-            .title("tiled + ggez example")
+            .title("rs-tiled + ggez example")
             .vsync(false)
         )
         .window_mode(ggez::conf::WindowMode::default()
             .dimensions(1000.0, 800.0)
         )
         .add_resource_path(""); // add repo root to ggez filesystem (our example map looks for `assets/tilesheet.png`)
+
     let (mut ctx, event_loop) = cb.build()?;
+
+    // construct and start the Game
     let state = Game::new(&mut ctx)?;
     event::run(ctx, event_loop, state)
 }
 
 struct Game {
-    map: Map,
-    tileset_image_cache: HashMap<String, graphics::Image>,
+    map: MapHandler,
     pan: (f32, f32),
     scale: f32,
-    animate: bool,
 }
 
 impl Game {
@@ -34,169 +37,13 @@ impl Game {
         let mut cache = FilesystemResourceCache::new();
         let map = Map::parse_file("assets/tiled_base64_external.tmx", &mut cache).unwrap();
 
-        // load images for the map's tilesets
-        let mut tileset_image_cache = HashMap::new();
-        for ts in map.tilesets().iter() {
-            if let Some(image) = &ts.image {
-                // image path comes in as "assets/tilesheet.png"
-                // but ggez needs it like "/assets/tilesheet.png" or it will complain
-                let mut pb: PathBuf = PathBuf::new();
-                pb.push("/");
-                pb.push(image.source.clone());
-                
-                // load the image
-                let mut img = graphics::Image::new(ctx, pb)?;
-                img.set_filter(graphics::FilterMode::Nearest);
-
-                tileset_image_cache.insert(
-                    ts.name.clone(),
-                    img,
-                );
-            }
-        }
+        let map_handler = MapHandler::new(map, ctx).unwrap();
 
         Ok(Self {
-            map,
-            tileset_image_cache,
+            map: map_handler,
             pan: (0.0, 0.0),
             scale: 1.0,
-            animate: false,
         })
-    }
-
-    fn generate_map_render(&mut self, ctx: &Context) -> Vec<Vec<SpriteBatch>> {
-        let mut layer_batches: Vec<Vec<SpriteBatch>> = Vec::new();
-
-        let tile_layers = self.map.layers().filter_map(|l| {
-            match l.layer_type() {
-                tiled::LayerType::TileLayer(tl) => Some((l.data(), tl)),
-                _ => None,
-            }
-        });
-
-        for (i, (layer, tl)) in tile_layers.enumerate() {
-            match &tl {
-                TileLayer::Finite(d) => {
-                    // create a sprite batch for each tileset
-                    // this needs to be done per layer otherwise the depth will be wrong when using tilesets on multiple layers
-                    let mut ts_batches = HashMap::new();
-                    let mut ts_sizes = HashMap::new();
-                    for ts in self.map.tilesets().iter() {
-                        if let Some(img) = self.tileset_image_cache.get(&ts.name) {
-                            // img.clone() here is cheap (see docs for `ggez::graphics::Image`)
-                            let batch = SpriteBatch::new(img.clone());
-                            ts_batches.insert(ts.name.clone(), batch);
-                            ts_sizes.insert(ts.name.clone(), (img.width(), img.height()));
-                        }
-                    }
-                    
-                    let width = d.data().width();
-                    let height = d.data().height();
-
-                    for x in 0..width as i32 {
-                        for y in 0..height as i32 {
-                            if let Some(tile) = d.get_tile(x, y) {
-                                let rect = get_tile_rect(tile.tileset, tile.id);
-                                if let Some(rect) = rect {
-                                    let ts = tile.tileset;
-                                    if let Some(ts_size) = ts_sizes.get(&ts.name) {
-                                        let mut dx = x as f32 * self.map.tile_width as f32 + self.pan.0 * (layer.parallax_x - 1.0);
-                                        let mut dy = y as f32 * self.map.tile_height as f32 + self.pan.1 * (layer.parallax_y - 1.0);
-
-                                        if self.animate {
-                                            dx += (ggez::timer::time_since_start(ctx).as_secs_f32() - x as f32 * 0.3 + i as f32 * 0.25).sin() * 20.0;
-                                            dy += (ggez::timer::time_since_start(ctx).as_secs_f32() * 1.25 + y as f32 * 0.3 + i as f32 * 0.25).cos() * 20.0;
-                                        }
-
-                                        let b = ts_batches.get_mut(&ts.name).unwrap();
-                                        b.add(
-                                            DrawParam::default()
-                                                .src(ggez::graphics::Rect::new(
-                                                    rect.0 as f32 / (*ts_size).0 as f32,
-                                                    rect.1 as f32 / (*ts_size).1 as f32,
-                                                    rect.2 as f32 / (*ts_size).0 as f32,
-                                                    rect.3 as f32 / (*ts_size).1 as f32,
-                                                ))
-                                                .dest([
-                                                    dx,
-                                                    dy,
-                                                ])
-                                                .color(ggez::graphics::Color::from_rgba(
-                                                    0xFF,
-                                                    0xFF,
-                                                    0xFF,
-                                                    (layer.opacity * 255.0) as u8,
-                                                )),
-                                        );
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    layer_batches.push(ts_batches.into_values().collect());
-                }
-                TileLayer::Infinite(_) => {
-                    unimplemented!()
-                }
-            }
-        }
-
-        layer_batches
-    }
-
-    fn draw_object(object: &tiled::ObjectData, ctx: &mut Context, draw_param: DrawParam) -> GameResult {
-        match &object.shape {
-            tiled::ObjectShape::Rect { width, height } => {
-                let bounds = graphics::Rect::new(object.x, object.y, *width, *height);
-                let shape =
-                    graphics::Mesh::new_rectangle(ctx, graphics::DrawMode::stroke(2.0), bounds, graphics::Color::CYAN)?;
-                graphics::draw(ctx, &shape, draw_param)?;
-            },
-            tiled::ObjectShape::Ellipse { width, height } => {
-                let shape = graphics::Mesh::new_ellipse(
-                    ctx, 
-                    graphics::DrawMode::stroke(2.0), 
-                    [object.x + width / 2.0, object.y + height / 2.0], 
-                    *width / 2.0, 
-                    *height / 2.0, 
-                    0.5, 
-                    graphics::Color::CYAN
-                )?;
-                graphics::draw(ctx, &shape, draw_param)?;
-            },
-            tiled::ObjectShape::Polyline { points } => {
-                let points: Vec<_> = points.iter().map(|p| [p.0 + object.x, p.1 + object.y]).collect();
-                let shape = graphics::Mesh::new_polyline(
-                    ctx, 
-                    graphics::DrawMode::stroke(2.0), 
-                    &points,
-                    graphics::Color::CYAN
-                )?;
-                graphics::draw(ctx, &shape, draw_param)?;
-            },
-            tiled::ObjectShape::Polygon { points } => {
-                let points: Vec<_> = points.iter().map(|p| [p.0 + object.x, p.1 + object.y]).collect();
-                let shape = graphics::Mesh::new_polyline(
-                    ctx, 
-                    graphics::DrawMode::stroke(2.0),
-                    &points,
-                    graphics::Color::CYAN
-                )?;
-                graphics::draw(ctx, &shape, draw_param)?;
-            },
-            tiled::ObjectShape::Point(_, _) => {
-                // exercise for the reader
-            },
-        }
-
-        if !object.name.is_empty() {
-            let text = graphics::Text::new(object.name.clone());
-            graphics::queue_text(ctx, &text, [object.x, object.y], Some(graphics::Color::YELLOW));
-            graphics::draw_queued_text(ctx, draw_param, None, graphics::FilterMode::Nearest)?;
-        }
-
-        Ok(())
     }
 }
 
@@ -206,47 +53,26 @@ impl event::EventHandler<ggez::GameError> for Game {
     }
 
     fn draw(&mut self, ctx: &mut Context) -> GameResult {
-        let bg_color: ggez::graphics::Color = self.map.background_color
-            .map(|c| ggez::graphics::Color::from_rgba(c.red, c.green, c.blue, 0xFF))
-            .unwrap_or([0.1, 0.2, 0.3, 1.0].into());
-
+        // fill background color
+        let bg_color: ggez::graphics::Color = self.map.background_color().unwrap_or([0.1, 0.2, 0.3, 1.0].into());
         graphics::clear(ctx, bg_color);
+
+        // draw map tiles + objects
 
         let window_size = graphics::size(ctx);
 
-        let layer_batches: Vec<Vec<SpriteBatch>> = self.generate_map_render(ctx);
-
         let draw_param = DrawParam::default()
             .dest([
-                self.pan.0 + window_size.0 / 2.0 - (self.map.width * self.map.tile_width) as f32 / 2.0, 
-                self.pan.1 + window_size.1 / 2.0 - (self.map.height * self.map.tile_height) as f32 / 2.0
+                self.pan.0 + window_size.0 / 2.0 - (self.map.width() * self.map.tile_width()) as f32 / 2.0, 
+                self.pan.1 + window_size.1 / 2.0 - (self.map.height() * self.map.tile_height()) as f32 / 2.0
                 ])
             .scale([self.scale, self.scale]);
 
-        // draw tile layers
-
-        // for each layer
-        for layer in &layer_batches {
-            // for each tileset in the layer
-            for batch in layer {
-                graphics::draw(ctx, batch, draw_param)?;
-            }
-        }
-
-        for l in self.map.layers() {
-            match &l.layer_type() {
-                tiled::LayerType::ObjectLayer(ol) => {
-                    for o in &ol.data().objects {
-                        Self::draw_object(o, ctx, draw_param.clone())?;
-                    }
-                }
-                _ => {},
-            }
-        }
+        self.map.draw(ctx, draw_param, self.pan)?;
 
         // draw map bounds
 
-        let rect = graphics::Rect::new(0.0, 0.0, (self.map.height * self.map.tile_height) as f32, (self.map.height * self.map.tile_height) as f32);
+        let rect = self.map.get_bounds();
         let r1 =
             graphics::Mesh::new_rectangle(ctx, graphics::DrawMode::stroke(2.0 / self.scale), rect, graphics::Color::from_rgb_u32(0x888888))?;
         graphics::draw(ctx, &r1, draw_param)?;
@@ -273,12 +99,14 @@ impl event::EventHandler<ggez::GameError> for Game {
     }
 
     fn mouse_button_down_event(&mut self, _ctx: &mut Context, button: event::MouseButton, _x: f32, _y: f32) {
+        // right click toggles demo animation effect
         if button == MouseButton::Right {
-            self.animate = !self.animate;
+            self.map.example_animate = !self.map.example_animate;
         }
     }
 
     fn mouse_motion_event(&mut self, ctx: &mut Context, _x: f32, _y: f32, dx: f32, dy: f32) {
+        // middle click + drag pans the map around
         if input::mouse::button_pressed(ctx, event::MouseButton::Middle) {
             self.pan.0 += dx;
             self.pan.1 += dy;
@@ -286,25 +114,17 @@ impl event::EventHandler<ggez::GameError> for Game {
     }
 
     fn mouse_wheel_event(&mut self, ctx: &mut Context, _x: f32, y: f32) {
+        // scroll wheel zooms
+
         let old_scale = self.scale;
         self.scale *= 1.0 + y as f32 * 0.1;
 
         // zoom to mouse cursor
         let mouse_pos = input::mouse::position(ctx);
         let window_size = graphics::size(ctx);
-        let pos_x = mouse_pos.x - window_size.0 / 2.0 + (self.map.width * self.map.tile_width) as f32 / 2.0;
-        let pos_y = mouse_pos.y - window_size.1 / 2.0 + (self.map.height * self.map.tile_height) as f32 / 2.0;
+        let pos_x = mouse_pos.x - window_size.0 / 2.0 + (self.map.width() * self.map.tile_width()) as f32 / 2.0;
+        let pos_y = mouse_pos.y - window_size.1 / 2.0 + (self.map.height() * self.map.tile_height()) as f32 / 2.0;
         self.pan.0 = (self.pan.0 - pos_x) / old_scale * self.scale + pos_x;
         self.pan.1 = (self.pan.1 - pos_y) / old_scale * self.scale + pos_y;
     }
-}
-
-fn get_tile_rect(tileset: &Tileset, id: u32) -> Option<(u32, u32, u32, u32)> {
-    let ts_x = id % tileset.columns;
-    let ts_y = id / tileset.columns;
-
-    let x = tileset.margin + (tileset.tile_width + tileset.spacing) * ts_x;
-    let y = tileset.margin + (tileset.tile_height + tileset.spacing) * ts_y;
-
-    Some((x, y, tileset.tile_width, tileset.tile_height))
 }
